Document Object Model
These files are about Document Object Model and corresponding changes in HTML. 
"The W3C Document Object Model (DOM) is a platform and language-neutral interface that allows programs and scripts to dynamically access and update the content, structure, and style of a document.
jQuery introduction also included that eases the use of JavaScript on the website. jQuery takes a lot of common tasks that require many lines of JavaScript code to accomplish, and wraps them into methods that you can call with a single line of code.